requirements.txt
anthropic
requests
python-dotenv
google-analytics-data
streamlit
